mlproject/requirements.txt
sentence-transformers==2.2.2
django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
numpy==1.24.3